권한체크

+++ src/main/java/com/koreait/surl_project_11/domain/surl/surl/controller/ApiV1SurlController.java
@@ -76,6 +76,12 @@ public class ApiV1SurlController {
     ) {
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
 
+        Member member = rq.getMember();
+
+        if (!surl.getAuthor().equals(member)) {
+            throw new GlobalException("403-1", "권한이 없어");
+        }
+
         return RsData.of(
                 new SurlGetRespBody(
                         new SurlDto(surl)
@@ -115,6 +121,12 @@ public class ApiV1SurlController {
     ) {
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
 
+        Member member = rq.getMember();
+
+        if (!surl.getAuthor().equals(member)) {
+            throw new GlobalException("403-1", "권한이 없어");
+        }
+
         surlService.delete(surl);
 
         return RsData.OK;
@@ -143,10 +155,19 @@ public class ApiV1SurlController {
     ) {
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
 
+        Member member = rq.getMember();
+
+        if (!surl.getAuthor().equals(member)) {
+//        if (surl.getAuthor().getId() != member.getId()) {
+            throw new GlobalException("403-1", "권한이 없어");
+        }
+
         RsData<Surl> modifyRs = surlService.modify(surl, reqBody.body, reqBody.url);
 
         return modifyRs.newDataOf(
-                new SurlModifyRespBody(
+                new
+
+                        SurlModifyRespBody(
                         new SurlDto(modifyRs.getData())
                 )
         );
